+++ HaselDebug/Tabs/UnlocksTab.Achievements.cs
@@ -1,14 +1,12 @@
-using System.Numerics;
-using Dalamud.Interface.Utility;
 using Dalamud.Interface.Utility.Raii;
 using FFXIVClientStructs.FFXIV.Client.UI.Agent;
 using HaselCommon.Extensions.Strings;
 using HaselCommon.Graphics;
-using HaselCommon.Gui;
 using HaselCommon.Services;
 using HaselDebug.Abstracts;
 using HaselDebug.Interfaces;
 using HaselDebug.Services;
+using HaselDebug.Utils;
 using ImGuiNET;
 using Achievement = FFXIVClientStructs.FFXIV.Client.Game.UI.Achievement;
 using AchievementSheet = Lumina.Excel.Sheets.Achievement;
@@ -18,7 +16,7 @@ namespace HaselDebug.Tabs;
 public unsafe class UnlocksTabAchievements(
     DebugRenderer DebugRenderer,
     ExcelService ExcelService,
-    TextureService TextureService) : DebugTab, ISubTab<UnlocksTab>
+    UnlocksTabUtils UnlocksTabUtils) : DebugTab, ISubTab<UnlocksTab>
 {
     private bool _achievementsRequested;
     private bool _achievementsHideSpoilers = true;
@@ -114,36 +112,11 @@ public unsafe class UnlocksTabAchievements(
             if (ImGui.IsItemHovered())
             {
                 if (canClick)
+                {
                     ImGui.SetMouseCursor(ImGuiMouseCursor.Hand);
+                }
 
-                using var tooltip = ImRaii.Tooltip();
-                if (!tooltip) continue;
-
-                using var popuptable = ImRaii.Table("PopupTable", 2, ImGuiTableFlags.BordersInnerV | ImGuiTableFlags.NoSavedSettings | ImGuiTableFlags.NoKeepColumnsVisible);
-                if (!popuptable) continue;
-
-                ImGui.TableSetupColumn("Icon", ImGuiTableColumnFlags.WidthFixed, 40 + ImGui.GetStyle().ItemInnerSpacing.X);
-                ImGui.TableSetupColumn("Text", ImGuiTableColumnFlags.WidthFixed, 300);
-
-                ImGui.TableNextColumn(); // Icon
-                TextureService.DrawIcon(row.Icon, 40);
-
-                ImGui.TableNextColumn(); // Text
-                using var indentSpacing = ImRaii.PushStyle(ImGuiStyleVar.IndentSpacing, ImGui.GetStyle().ItemInnerSpacing.X);
-                using var indent = ImRaii.PushIndent(1);
-
-                ImGui.TextUnformatted(name);
-                ImGuiUtils.PushCursorY(-3);
-                using (ImRaii.PushColor(ImGuiCol.Text, (uint)Color.Grey))
-                    ImGui.TextUnformatted(categoryName);
-                ImGuiUtils.PushCursorY(1);
-
-                // separator
-                var pos = ImGui.GetCursorScreenPos();
-                ImGui.GetWindowDrawList().AddLine(pos, pos + new Vector2(ImGui.GetContentRegionAvail().X, 0), ImGui.GetColorU32(ImGuiCol.Separator));
-                ImGuiUtils.PushCursorY(4);
-
-                ImGuiHelpers.SafeTextWrapped(description);
+                UnlocksTabUtils.DrawTooltip((uint)row.Icon, name, categoryName, description);
             }
         }
     }