+++ HaselDebug/Tabs/UnlocksTab.Minions.cs
@@ -1,16 +1,14 @@
-using System.Numerics;
-using Dalamud.Interface.Utility;
 using Dalamud.Interface.Utility.Raii;
 using FFXIVClientStructs.FFXIV.Client.Game;
 using FFXIVClientStructs.FFXIV.Client.Game.Control;
 using FFXIVClientStructs.FFXIV.Client.Game.UI;
 using HaselCommon.Extensions.Strings;
 using HaselCommon.Graphics;
-using HaselCommon.Gui;
 using HaselCommon.Services;
 using HaselDebug.Abstracts;
 using HaselDebug.Interfaces;
 using HaselDebug.Services;
+using HaselDebug.Utils;
 using ImGuiNET;
 using Lumina.Excel.Sheets;
 
@@ -20,7 +18,8 @@ public unsafe class UnlocksTabMinions(
     DebugRenderer DebugRenderer,
     ExcelService ExcelService,
     TextService TextService,
-    TextureService TextureService) : DebugTab, ISubTab<UnlocksTab>
+    TextureService TextureService,
+    UnlocksTabUtils UnlocksTabUtils) : DebugTab, ISubTab<UnlocksTab>
 {
     public override string Title => "Minions";
     public override bool DrawInChild => false;
@@ -73,7 +72,7 @@ public unsafe class UnlocksTabMinions(
                 }
                 else
                 {
-                    ImGui.TextUnformatted(name);
+                    ImGui.Selectable(name);
                 }
             }
 
@@ -82,33 +81,13 @@ public unsafe class UnlocksTabMinions(
                 if (canUse)
                     ImGui.SetMouseCursor(ImGuiMouseCursor.Hand);
 
-                using var tooltip = ImRaii.Tooltip();
-                if (!tooltip) continue;
-
-                using var popuptable = ImRaii.Table("PopupTable", 2, ImGuiTableFlags.BordersInnerV | ImGuiTableFlags.NoSavedSettings | ImGuiTableFlags.NoKeepColumnsVisible);
-                if (!popuptable) continue;
-
-                ImGui.TableSetupColumn("Icon", ImGuiTableColumnFlags.WidthFixed, 40 + ImGui.GetStyle().ItemInnerSpacing.X);
-                ImGui.TableSetupColumn("Text", ImGuiTableColumnFlags.WidthFixed, 300);
-
-                ImGui.TableNextColumn(); // Icon
-                TextureService.DrawIcon(row.Icon, 40);
-
-                ImGui.TableNextColumn(); // Text
-                using var indentSpacing = ImRaii.PushStyle(ImGuiStyleVar.IndentSpacing, ImGui.GetStyle().ItemInnerSpacing.X);
-                using var indent = ImRaii.PushIndent(1);
-
-                ImGui.TextUnformatted(name);
-
-                if (ExcelService.TryGetRow<CompanionTransient>(row.RowId, out var transient))
-                {
-                    // separator
-                    var pos = ImGui.GetCursorScreenPos();
-                    ImGui.GetWindowDrawList().AddLine(pos, pos + new Vector2(ImGui.GetContentRegionAvail().X, 0), ImGui.GetColorU32(ImGuiCol.Separator));
-                    ImGuiUtils.PushCursorY(4);
-
-                    ImGuiHelpers.SafeTextWrapped(transient.DescriptionEnhanced.ExtractText().StripSoftHypen());
-                }
+                UnlocksTabUtils.DrawTooltip(
+                    row.Icon,
+                    name,
+                    null,
+                    ExcelService.TryGetRow<CompanionTransient>(row.RowId, out var transient) && !transient.DescriptionEnhanced.IsEmpty
+                        ? transient.DescriptionEnhanced.ExtractText().StripSoftHypen()
+                        : null);
             }
         }
     }